emotion-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
deepface>=0.0.93
tf-keras>=2.16.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
modal>=0.67.0
